Add files via upload
Smart Parking is a parking strategy that combines technology and human innovation in an effort to use as few resources as possible—such as fuel, time and space—to achieve faster, easier and denser parking of vehicles for the majority of time they remain idle. A vehicle parking system that helps drivers find a vacant spot. Using sensors in each parking space that detect the presence or absence of a vehicle, signs direct incoming drivers to available locations.